requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
ultralytics==8.0.196
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
python-multipart==0.0.6
channels==4.0.0
channels-redis==4.1.0
websockets==11.0.3
redis==5.0.1
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
dj-database-url==2.1.0 